+++ src/FakeXrmEasy.Core/CommercialLicense/SubscriptionValidator.cs
@@ -88,7 +88,7 @@ namespace FakeXrmEasy.Core.CommercialLicense
 
         internal bool IsUsageValid()
         {
-            if (IsRunningInContinuousIntegration())
+            if (_environmentReader.IsRunningInContinuousIntegration())
             {
                 return true;
             }
@@ -117,12 +117,6 @@ namespace FakeXrmEasy.Core.CommercialLicense
                 
             }
             return true;
-        }
-
-        private bool IsRunningInContinuousIntegration()
-        {
-            return "1".Equals(_environmentReader.GetEnvironmentVariable("FAKE_XRM_EASY_CI"))
-                || "True".Equals(_environmentReader.GetEnvironmentVariable("TF_BUILD"));
         }
     }
 }